tests: move gossip simulator into tests module
